+++ test/config_test.rb
@@ -78,6 +78,7 @@ class ConfigTest < Minitest::Test
         jshint: nil,
         ktlint: nil,
         languagetool: nil,
+        metrics_complexity: nil,
         misspell: nil,
         phinder: nil,
         phpmd: nil,