+++ test/smokes/metrics_complexity/comma_in_filename/src/example,file.cs
@@ -1,0 +1,15 @@
+/* test file */
+
+public class Program
+{
+  public static int TestFunction(string a = "foo", string b = "", double c = 3.14)
+  {
+    /* test function */
+    if (a == "bar")
+    {
+      return 0;
+    }
+
+    return 1;
+  }
+}