+++ test/smokes/metrics_complexity/root_dir/src/baz/qux/今日は世界.go
@@ -1,0 +1,12 @@
+
+package main
+import "fmt"
+
+func main() {
+  hello_world()
+}
+
+func hello_world() {
+/* Hello world in Go */
+  fmt.Printf("Hello, World!\n")
+}