+++ test/smokes/metrics_complexity/root_dir/src/baz/こんにちは.py
@@ -1,0 +1,5 @@
+def hello_world():
+# Hello world in Python
+  print("Hello, World!")
+
+hello_world()