+++ test/smokes/metrics_complexity/success/src/baz/hello.php
@@ -1,0 +1,10 @@
+<?php
+function hello_world()
+{
+/* Hello world in PHP */
+  echo 'Hello, World!';
+}
+
+hello_world();
+?>
+