+++ test/smokes/metrics_complexity/success/src/baz/hello.scala
@@ -1,0 +1,9 @@
+object HelloWorld extends App {
+  def hello_world() {
+/* Hello world in Scala */
+    println("Hello, World!")
+  }
+
+  hello_world()
+}
+