+++ test/smokes/metrics_complexity/success/src/foo/bar/hello.m
@@ -1,0 +1,14 @@
+#import <Foundation/Foundation.h>
+
+static void hello_world() {
+/* Hello world in Objective-C */
+  @autoreleasepool {
+    NSLog(@"Hello, World!");
+  }
+}
+
+int main(int argc, const char * argv[]) {
+  hello_world();
+  return 0;
+}
+