+++ test/smokes/metrics_complexity/success/src/foo/hello.cpp
@@ -1,0 +1,17 @@
+#include <iostream>
+using namespace std;
+
+void print_hello_world();
+
+int main(void)
+{
+  print_hello_world();
+  return 0;
+}
+
+void print_hello_world()
+{
+/* Hello world in C++ */
+  cout << "Hello, World!" << endl;
+}
+