+++ test/smokes/metrics_complexity/success/src/foo/hello.java
@@ -1,0 +1,11 @@
+class HelloWorld {
+  static public void main(String args[]) {
+    printHelloWorld();
+  }
+
+  static public void printHelloWorld() {
+/* Hello world in Java */
+    System.out.println("Hello, World!");
+  }
+}
+