+++ test/smokes/metrics_complexity/success/src/foo/hello.js
@@ -1,0 +1,7 @@
+function hello_world() {
+/* Hello world in JavaScript */
+  console.log("Hello, World!");
+}
+
+hello_world();
+